Extracts ClusterTests
